Fixed mask array bug
Because an array was initialized with zero values, when an x value was
skipped due to being masked, the count in the zeroth bin was still
incremented. Fixed by initializing this array with a value > the number
of bins, so the masked data is properly ignored.

This code will need refactored, it is very slow on big datasets.